+++ libopendmi/src/entity/cooling-device.c
@@ -79,7 +79,7 @@ const dmi_entity_spec_t dmi_cooling_device_spec =
     .name            = "Cooling device",
     .type            = DMI_TYPE_COOLING_DEVICE,
     .minimum_version = DMI_VERSION(2, 2, 0),
-    .minimum_length  = 0x0E,
+    .minimum_length  = 0x0C,
     .decoded_length  = sizeof(dmi_cooling_device_t),
     .attributes      = (dmi_attribute_t[]){
         DMI_ATTRIBUTE(dmi_cooling_device_t, probe_handle, HANDLE, {