+++ CHANGELOG.md
@@ -11,12 +11,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added support for enumeration range names
 - Moved `<opendmi/name.h>` to `<opendmi/utils/name.h>`
 - Refactored entity decoders API
+- Fixed group associations items list decoding
 
 ## [0.1.4] - Unreleased
 
+- Fixed cooling device structure minimum length
+- Fixed current probe structure minimum length
+- Fixed temperature probe structure mimimum length
+- Fixed voltage probe structure minimum length
 - Fixed hardware security settings decoding
 - Fixed baseboard structure minumum length
-- Fixed group associations items list decoding
 
 ## [0.1.3] - January 11, 2026